server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pydantic==2.11.7
pydantic_core==2.33.2
regex==2025.7.34
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
ai==3.1.3
soundfile==0.12.1
elevenlabs==1.2.0
# Added for VAD, Diarization, and Speaker Verification
silero-vad==0.0.1
speechbrain==0.5.16
torchaudio==2.2.0 # Required by speechbrain
